Fixes for 1.73.0

+++ recipe/install-py-boost.sh
@@ -27,13 +27,8 @@ done
 # instead of boost_python3. It also does not add any specified
 # --python-buildid; ping @stefanseefeld
 pushd "${PREFIX}/lib"
-  ln -s libboost_python${PY_VER//./}.a libboost_python.a
-  ln -s libboost_numpy${PY_VER//./}.a libboost_numpy.a
-  if [[ ${target_platform} == osx-64 ]]; then
-    ln -s libboost_python${PY_VER//./}.dylib libboost_python.dylib
-    ln -s libboost_numpy${PY_VER//./}.dylib libboost_numpy.dylib
-  else
-    ln -s libboost_python${PY_VER//./}.so libboost_python.so
-    ln -s libboost_numpy${PY_VER//./}.so libboost_numpy.so
-  fi
+  [[ -f libboost_python.a ]] || ln -s libboost_python${PY_VER//./}.a libboost_python.a
+  [[ -f libboost_numpy.a ]] || ln -s libboost_numpy${PY_VER//./}.a libboost_numpy.a
+  ln -s libboost_python${PY_VER//./}${SHLIB_EXT} libboost_python${SHLIB_EXT}
+  ln -s libboost_numpy${PY_VER//./}${SHLIB_EXT} libboost_numpy${SHLIB_EXT}
 popd
